refactor: move graasp schema, small fixes
