+++ package.json
@@ -116,7 +116,7 @@
     "start": "concurrently \"yarn start:dev\" \"wait-on http://localhost:3000 && env-cmd -f ./.env.local electron .\"",
     "start:dev": "env-cmd -f ./.env.local react-scripts start",
     "premocha": "yarn prepare",
-    "mocha": "concurrently \"mkdirp test/tmp\" \"env-cmd -f ./.env.test react-scripts start\" \"wait-on http://localhost:3000 && mocha --recursive  --require @babel/register test\"",
+    "mocha": "concurrently \"mkdirp test/tmp\" \"env-cmd -f ./.env.test react-scripts start\" \"wait-on http://localhost:3000 && mocha --recursive --retries 3 --require @babel/register test\"",
     "mocha:dev": "mocha --recursive --require @babel/register test",
     "prettier:check": "prettier --check {src,public,test}/**{,/**}/*.js",
     "prettier:write": "prettier --write {src,public,test}/**{,/**}/*.js",
@@ -126,7 +126,7 @@
     "release:tags": "git fetch --tags && git add CHANGELOG.md && standard-version -a && git push --follow-tags origin master",
     "release": "yarn release:tags && env-cmd -f ./.env electron-builder -ml && env-cmd -f ./.env electron-builder -w --x64 --ia32",
     "ci": "react-scripts test --env=jsdom --coverage",
-    "prepare": "mkdirp public/shared && cpy \"src/shared/**/*.js\" \"public/shared\"",
+    "prepare": "mkdirp public/shared && cpy \"./**/*\" \"../../public/shared\" --parents --cwd=src/shared",
     "setup": "node scripts/setup.js",
     "test:once": "cross-env CI=true env-cmd -f ./.env.test react-scripts test --env=jsdom",
     "test": "react-scripts test"
