+++ public/shared/channels.js
@@ -32,6 +32,7 @@ module.exports = {
   SAVE_ALGORITHM_CHANNEL: 'algorithm:save',
   ADD_ALGORITHM_CHANNEL: 'algorithm:add',
   ADD_DEFAULT_ALGORITHM_CHANNEL: 'algorithm:add:default',
+  GET_ALGORITHM_CODE_CHANNEL: 'algorithm:get:code',
   BROWSE_FILE_CHANNEL: 'file:browse',
   GET_UTILS_CHANNEL: 'utils:get',
   SAVE_UTILS_CHANNEL: 'utils:save',
@@ -50,5 +51,4 @@ module.exports = {
   SET_SCHEMA_CHANNEL: 'schema:set',
   DELETE_SCHEMA_CHANNEL: 'schema:delete',
   OPEN_URL_IN_BROWSER_CHANNEL: 'open:browser',
-  GET_ALGORITHM_CODE_CHANNEL: 'algorithm:code:get',
 };
