+++ src/langs/en.json
@@ -43,7 +43,7 @@
     "Dataset": "Dataset",
     "Algorithm": "Algorithm",
     "Load a dataset first": "Load a dataset first",
-    "No algorithms are available": "No algorithms are available",
+    "No algorithms available": "No algorithms available",
     "Execute": "Execute",
     "Executions": "Executions",
     "Author": "Author",
