+++ src/langs/fr.json
@@ -43,7 +43,7 @@
     "Dataset": "Dataset",
     "Algorithm": "Algorithme",
     "Load a dataset first": "Veuillez importer un dataset",
-    "No algorithms are available": "Aucun algorithme n'est disponible",
+    "No algorithms available": "Aucun algorithme disponible",
     "Execute": "Exécuter",
     "Executions": "Exécutions",
     "Author": "Auteur",
